requirements.txt
# Database drivers
psycopg2-binary>=2.9.9
pymongo>=4.6.1
couchdb>=1.2
cassandra-driver>=3.29.0

# Data processing and visualization
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
tabulate>=0.9.0

# Utilities
tqdm>=4.66.1
python-dotenv>=1.0.0
requests>=2.31.0
click>=8.1.7

# Performance monitoring
psutil>=5.9.7
memory-profiler>=0.61.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1 